Refactor timer logic to pass FCC tests
My solution relied on requestAnimationFrame and Date.now to calculate
the elapsed time on each frame.  This approach is not supported by the
FCC tests for this project, which assume the use of setInterval or
setTimeout.

I tweaked code to use setInterval istead of RAF (and disabled the
progress animation)